did all the things

+++ db/migrate/11_add_fte_to_role.rb
@@ -1,5 +1,5 @@
 class AddFteToRole < ActiveRecord::Migration
   def change
-    add_column :roles, :fte, :integer, null: false
+    add_column :roles, :fte, :float, null: false
   end
 end
